requirements.txt
boto3==1.1.3
botocore==1.2.0
certifi==2015.9.6.1
docutils==0.12
elasticsearch==1.6.0
futures==2.2.0
jmespath==0.7.1
nose==1.3.7
pkginfo==1.2.1
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
twine==1.5.0
urllib3==1.10.4
wheel==0.24.0
